refactor: snapshot creator and timeframe handler
This streamlines the snapshot creation logic, and allows for other types of logic to be executed when timeframe changes are detected.
