+++ src/common/mod.rs
@@ -1,5 +1,6 @@
 pub mod conversion;
 pub mod event_extraction;
 pub mod format;
+pub mod pool_utils;
 pub mod prices;
 pub mod utils;
