+++ src/lib.rs
@@ -1,12 +1,13 @@
 mod abi;
+mod common;
 mod constants;
+mod snapshot;
+mod key_management;
 mod modules;
 mod network_config;
 mod pb;
 mod rpc;
-mod key_management;
 mod types;
-mod common;
 
 pub use modules::*;
