+++ src/modules/420_graph_out.rs
@@ -1,8 +1,4 @@
-use std::{
-    borrow::BorrowMut,
-    collections::HashSet,
-    ops::{Div, Mul},
-};
+use std::{cell::RefCell, collections::HashSet, rc::Rc};
 
 use anyhow::anyhow;
 use substreams::{
@@ -19,7 +15,7 @@ use substreams_entity_change::{pb::entity::EntityChanges, tables::Tables};
 use crate::{
     common::{
         format::{self, format_address_string},
-        prices::get_token_usd_price,
+        pool_utils::{get_input_token_balances, get_input_token_weights},
         utils,
     },
     constants,
@@ -34,6 +30,11 @@ use crate::{
             Events, Pool, PoolFee, PoolFees, Pools, Token,
         },
         uniswap_pricing::v1::Erc20Price,
+    },
+    snapshot::{
+        snapshot_creator::SnapshotCreator,
+        timeframe_change_handler::TimeframeChangeHandler,
+        utils::{prepare_snapshot_closure, separate_timeframe_deltas},
     },
     types::snapshot::SnapshotType,
 };
@@ -62,11 +63,12 @@ pub fn graph_out(
     pool_tvl_store: StoreGetBigDecimal,
     pool_tvl_deltas: Deltas<DeltaBigDecimal>,
     protocol_tvl_store: StoreGetBigDecimal,
-    store_current_time_snapshots: Deltas<DeltaInt64>,
+    current_time_deltas: Deltas<DeltaInt64>,
     uniswap_prices: StoreGetProto<Erc20Price>,
     chainlink_prices: StoreGetBigDecimal,
 ) -> Result<EntityChanges, Error> {
     let mut tables = Tables::new();
+
     create_protocol_entity(&mut tables, &clock);
 
     // Create entities related to Pool contract deployments
@@ -153,341 +155,80 @@ pub fn graph_out(
         &protocol_tvl_store,
     );
 
-    for delta in store_current_time_snapshots.deltas {
-        if delta.operation == Operation::Create {
-            continue;
-        }
-        // This condition indicates the delta relates to a new day.
-        if delta.key == StoreKey::current_day_id_key()
-            && delta.operation == Operation::Update
-            && delta.old_value != delta.new_value
-        {
-            let day_id = delta.old_value;
-            // We have identified a new day, create snapshots for the previous day.
-            create_protocol_financials_daily_snapshot(
-                &clock,
-                &mut tables,
-                &day_id,
-                &protocol_tvl_store,
-                &protocol_volume_store,
-            );
-            create_liquidity_pool_snapshots(
-                &SnapshotType::Daily,
-                &day_id,
-                &mut tables,
-                &clock,
-                &pool_count_store,
-                &pool_addresses_store,
-                &pools_store,
-                &pool_tvl_store,
-                &pool_volume_usd_store,
-                &pool_volume_native_store,
-                &input_token_balances_store,
-                &output_token_supply_store,
-                &uniswap_prices,
-                &chainlink_prices,
-            );
-        }
-        // This condition indicates the delta relates to a new hour
-        else if delta.key == StoreKey::current_hour_id_key()
-            && delta.operation == Operation::Update
-            && delta.old_value != delta.new_value
-        {
-            let hour_id = delta.old_value;
-            create_liquidity_pool_snapshots(
-                &SnapshotType::Hourly,
-                &hour_id,
-                &mut tables,
-                &clock,
-                &pool_count_store,
-                &pool_addresses_store,
-                &pools_store,
-                &pool_tvl_store,
-                &pool_volume_usd_store,
-                &pool_volume_native_store,
-                &input_token_balances_store,
-                &output_token_supply_store,
-                &uniswap_prices,
-                &chainlink_prices,
-            );
-        }
-    }
+    process_timeframe_deltas(
+        &clock,
+        &current_time_deltas,
+        &mut tables,
+        &pool_count_store,
+        &pool_addresses_store,
+        &pools_store,
+        &pool_tvl_store,
+        &pool_volume_usd_store,
+        &pool_volume_native_store,
+        &protocol_tvl_store,
+        &protocol_volume_store,
+        &input_token_balances_store,
+        &output_token_supply_store,
+        &uniswap_prices,
+        &chainlink_prices,
+    );
 
     Ok(tables.to_entity_changes())
 }
 
-fn create_liquidity_pool_snapshots(
-    snapshot_type: &SnapshotType,
-    time_frame_id: &i64,
-    tables: &mut Tables,
+fn process_timeframe_deltas(
     clock: &Clock,
+    deltas: &Deltas<DeltaInt64>,
+    tables: &mut Tables,
     pool_count_store: &StoreGetInt64,
     pool_addresses_store: &StoreGetString,
     pools_store: &StoreGetProto<Pool>,
     pool_tvl_store: &StoreGetBigDecimal,
     pool_volume_usd_store: &StoreGetBigDecimal,
     pool_volume_native_store: &StoreGetBigInt,
+    protocol_tvl_store: &StoreGetBigDecimal,
+    protocol_volume_store: &StoreGetBigDecimal,
     input_token_balances_store: &StoreGetBigInt,
     output_token_supply_store: &StoreGetBigInt,
     uniswap_prices: &StoreGetProto<Erc20Price>,
     chainlink_prices: &StoreGetBigDecimal,
 ) {
-    // Create Liquidity Pools daily snapshots
-    let pool_count = pool_count_store
-        .get_last(StoreKey::protocol_pool_count_key())
-        .unwrap();
+    let snapshot_creator = Rc::new(RefCell::new(SnapshotCreator::new(
+        tables,
+        clock,
+        pool_count_store,
+        pool_addresses_store,
+        pools_store,
+        pool_tvl_store,
+        pool_volume_usd_store,
+        pool_volume_native_store,
+        protocol_tvl_store,
+        protocol_volume_store,
+        input_token_balances_store,
+        output_token_supply_store,
+        uniswap_prices,
+        chainlink_prices,
+    )));
 
-    for i in 1..=pool_count {
-        let pool_address = pool_addresses_store
-            .get_last(StoreKey::pool_address_key(&i))
-            .unwrap();
+    let (daily_deltas, hourly_deltas) = separate_timeframe_deltas(deltas);
 
-        let pool = pools_store
-            .get_last(StoreKey::pool_key(&pool_address))
-            .unwrap();
+    // Prepare closures for handling new day and new hour events.
+    // These closures will utilize the SnapshotCreator to generate snapshots.
+    // We use Rc::clone to ensure the SnapshotCreator can be shared among closures without taking ownership.
+    let on_new_day = prepare_snapshot_closure(Rc::clone(&snapshot_creator), SnapshotType::Daily);
+    let on_new_hour = prepare_snapshot_closure(Rc::clone(&snapshot_creator), SnapshotType::Hourly);
 
-        let pool_tvl_usd = pool_tvl_store
-            .get_last(StoreKey::pool_tvl_key(&pool.address))
-            .unwrap();
+    // Initialize the TimeframeChangeHandler with the separated deltas and the prepared closures.
+    // This handler will check for updates in daily and hourly deltas and trigger the appropriate closures.
+    let mut timeframe_change_handler = TimeframeChangeHandler {
+        daily_deltas: &daily_deltas,
+        hourly_deltas: &hourly_deltas,
+        on_new_day: Box::new(on_new_day),
+        on_new_hour: Box::new(on_new_hour),
+    };
 
-        // Get the volume in the pool for a given timeframe (Daily/Hourly)
-        let pool_volume = match snapshot_type {
-            SnapshotType::Daily => pool_volume_usd_store
-                .get_last(StoreKey::pool_volume_usd_daily_key(
-                    &pool.address,
-                    &time_frame_id,
-                ))
-                .unwrap(),
-            SnapshotType::Hourly => pool_volume_usd_store
-                .get_last(StoreKey::pool_volume_usd_hourly_key(
-                    &pool.address,
-                    &time_frame_id,
-                ))
-                .unwrap(),
-        };
-
-        // Get the volumes of each pool input token for a given timeframe (Daily/Hourly)
-        let (volume_by_token_native, volume_by_token_usd) = get_pool_token_volumes_in_timeframe(
-            &pool,
-            time_frame_id,
-            snapshot_type,
-            pool_volume_native_store,
-            pool_volume_usd_store,
-        );
-
-        let pool_cumulative_volume_usd = pool_volume_usd_store
-            .get_last(StoreKey::pool_volume_usd_key(&pool.address))
-            .unwrap();
-
-        let input_token_balances = get_input_token_balances(
-            &pool_address,
-            &pool.input_tokens,
-            &input_token_balances_store,
-        );
-
-        let input_token_weights = get_input_token_weights(&pool, &pool_tvl_store);
-
-        let output_token_supply = output_token_supply_store
-            .get_last(StoreKey::output_token_supply_key(&pool_address))
-            .unwrap_or_else(|| BigInt::zero());
-
-        let output_token_price =
-            get_token_usd_price(pool.output_token_ref(), &uniswap_prices, &chainlink_prices);
-
-        // Create the relevant timeframe snapshot
-        match snapshot_type {
-            SnapshotType::Daily => create_pool_daily_snapshot(
-                tables,
-                clock,
-                time_frame_id,
-                &pool_address,
-                &pool_tvl_usd,
-                &pool_volume,
-                &volume_by_token_native,
-                &volume_by_token_usd,
-                &pool_cumulative_volume_usd,
-                &input_token_balances,
-                &input_token_weights,
-                &output_token_supply,
-                &output_token_price,
-            ),
-            SnapshotType::Hourly => create_pool_hourly_snapshot(
-                tables,
-                clock,
-                time_frame_id,
-                &pool_address,
-                &pool_tvl_usd,
-                &pool_volume,
-                &volume_by_token_native,
-                &volume_by_token_usd,
-                &pool_cumulative_volume_usd,
-                &input_token_balances,
-                &input_token_weights,
-                &output_token_supply,
-                &output_token_price,
-            ),
-        }
-    }
-}
-
-fn create_pool_daily_snapshot(
-    tables: &mut Tables,
-    clock: &Clock,
-    day_id: &i64,
-    pool_address: &str,
-    pool_tvl_usd: &BigDecimal,
-    pool_volume_daily: &BigDecimal,
-    volume_by_token_native: &Vec<BigInt>,
-    volume_by_token_usd: &Vec<BigDecimal>,
-    pool_cumulative_volume_usd: &BigDecimal,
-    input_token_balances: &Vec<BigInt>,
-    input_token_weights: &Vec<BigDecimal>,
-    output_token_supply: &BigInt,
-    output_token_price: &BigDecimal,
-) {
-    tables
-        .create_row(
-            "LiquidityPoolDailySnapshot",
-            EntityKey::pool_daily_snapshot_key(pool_address, day_id),
-        )
-        .set("protocol", EntityKey::protocol_key())
-        .set("pool", EntityKey::liquidity_pool_key(pool_address))
-        .set("blockNumber", BigInt::from(clock.number))
-        .set(
-            "timestamp",
-            BigInt::from(clock.timestamp.clone().unwrap().seconds),
-        )
-        .set("totalValueLockedUSD", pool_tvl_usd)
-        .set("dailyVolumeUSD", pool_volume_daily)
-        .set("dailyVolumeByTokenAmount", volume_by_token_native)
-        .set("dailyVolumeByTokenUSD", volume_by_token_usd)
-        .set("cumulativeVolumeUSD", pool_cumulative_volume_usd)
-        .set("inputTokenBalances", input_token_balances)
-        .set("inputTokenWeights", input_token_weights)
-        .set("outputTokenSupply", output_token_supply)
-        .set("outputTokenPriceUSD", output_token_price);
-}
-
-fn create_pool_hourly_snapshot(
-    tables: &mut Tables,
-    clock: &Clock,
-    hour_id: &i64,
-    pool_address: &str,
-    pool_tvl_usd: &BigDecimal,
-    pool_volume_hourly: &BigDecimal,
-    volume_by_token_native: &Vec<BigInt>,
-    volume_by_token_usd: &Vec<BigDecimal>,
-    pool_cumulative_volume_usd: &BigDecimal,
-    input_token_balances: &Vec<BigInt>,
-    input_token_weights: &Vec<BigDecimal>,
-    output_token_supply: &BigInt,
-    output_token_price: &BigDecimal,
-) {
-    tables
-        .create_row(
-            "LiquidityPoolHourlySnapshot",
-            EntityKey::pool_hourly_snapshot_key(&pool_address, hour_id),
-        )
-        .set("protocol", EntityKey::protocol_key())
-        .set("pool", EntityKey::liquidity_pool_key(pool_address))
-        .set("blockNumber", BigInt::from(clock.number))
-        .set(
-            "timestamp",
-            BigInt::from(clock.timestamp.clone().unwrap().seconds),
-        )
-        .set("totalValueLockedUSD", pool_tvl_usd)
-        .set("hourlyVolumeUSD", pool_volume_hourly)
-        .set("hourlyVolumeByTokenAmount", volume_by_token_native)
-        .set("hourlyVolumeByTokenUSD", volume_by_token_usd)
-        .set("cumulativeVolumeUSD", pool_cumulative_volume_usd)
-        .set("inputTokenBalances", input_token_balances)
-        .set("inputTokenWeights", input_token_weights)
-        .set("outputTokenSupply", output_token_supply)
-        .set("outputTokenPriceUSD", output_token_price);
-}
-
-fn get_pool_token_volumes_in_timeframe(
-    pool: &Pool,
-    time_frame_id: &i64,
-    snapshot_type: &SnapshotType,
-    pool_volume_native_store: &StoreGetBigInt,
-    pool_volume_usd_store: &StoreGetBigDecimal,
-) -> (Vec<BigInt>, Vec<BigDecimal>) {
-    let mut pool_volume_by_token_native: Vec<BigInt> = Vec::new();
-    let mut pool_volume_by_token_usd: Vec<BigDecimal> = Vec::new();
-
-    for token in &pool.input_tokens {
-        let native_volume_key = match snapshot_type {
-            SnapshotType::Daily => StoreKey::pool_token_volume_native_daily_key(
-                &pool.address,
-                &token.address,
-                &time_frame_id,
-            ),
-            SnapshotType::Hourly => StoreKey::pool_token_volume_native_hourly_key(
-                &pool.address,
-                &token.address,
-                &time_frame_id,
-            ),
-        };
-        let usd_volume_key = match snapshot_type {
-            SnapshotType::Daily => StoreKey::pool_token_volume_usd_daily_key(
-                &pool.address,
-                &token.address,
-                &time_frame_id,
-            ),
-            SnapshotType::Hourly => StoreKey::pool_token_volume_usd_hourly_key(
-                &pool.address,
-                &token.address,
-                &time_frame_id,
-            ),
-        };
-
-        // Fetch and push the native volume for the token
-        let token_native_volume = pool_volume_native_store
-            .get_last(&native_volume_key)
-            .unwrap_or_else(|| BigInt::from(0)); // Default to 0 if not found
-        pool_volume_by_token_native.push(token_native_volume);
-
-        // Fetch and push the USD volume for the token
-        let token_usd_volume = pool_volume_usd_store
-            .get_last(&usd_volume_key)
-            .unwrap_or_else(|| BigDecimal::from(0)); // Default to 0 if not found
-        pool_volume_by_token_usd.push(token_usd_volume);
-    }
-
-    (pool_volume_by_token_native, pool_volume_by_token_usd)
-}
-
-fn create_protocol_financials_daily_snapshot(
-    clock: &Clock,
-    tables: &mut Tables,
-    day_id: &i64,
-    protocol_tvl_store: &StoreGetBigDecimal,
-    protocol_volume_store: &StoreGetBigDecimal,
-) {
-    let tvl_usd = protocol_tvl_store
-        .get_last(StoreKey::protocol_tvl_key())
-        .unwrap();
-    let daily_volume = protocol_volume_store
-        .get_last(StoreKey::protocol_daily_volume_usd_key(&day_id))
-        .unwrap();
-    let cumulative_volume = protocol_volume_store
-        .get_last(StoreKey::protocol_volume_usd_key())
-        .unwrap();
-    tables
-        .create_row(
-            "FinancialsDailySnapshot",
-            EntityKey::protocol_daily_financials_key(&day_id),
-        )
-        .set("protocol", EntityKey::protocol_key())
-        .set("totalValueLockedUSD", tvl_usd)
-        .set("dailyVolumeUSD", daily_volume)
-        .set("cumulativeVolumeUSD", cumulative_volume)
-        .set("blockNumber", BigInt::from(clock.number))
-        .set(
-            "timestamp",
-            BigInt::from(clock.timestamp.clone().unwrap().seconds),
-        );
+    // Process the timeframe changes by iterating over deltas and triggering closures if conditions are met.
+    timeframe_change_handler.handle_timeframe_changes();
 }
 
 fn create_protocol_entity(tables: &mut Tables, clock: &Clock) {
@@ -738,51 +479,6 @@ fn update_pool_output_token_supply(
             EntityKey::liquidity_pool_key(&event.pool_address),
         )
         .set("outputTokenSupply", output_token_supply);
-}
-
-fn get_input_token_balances(
-    pool_address: &str,
-    input_tokens: &Vec<Token>,
-    input_token_balances_store: &StoreGetBigInt,
-) -> Vec<BigInt> {
-    input_tokens
-        .iter()
-        .map(|token| {
-            let input_token_balance_key =
-                StoreKey::input_token_balance_key(&pool_address, &token.address);
-            input_token_balances_store
-                .get_last(input_token_balance_key)
-                .unwrap_or_else(|| {
-                    substreams::log::debug!(
-                        "No input token balance found for pool {} and token {}",
-                        pool_address,
-                        token.address
-                    );
-                    BigInt::zero()
-                })
-        })
-        .collect()
-}
-
-fn get_input_token_weights(pool: &Pool, pool_tvl_store: &StoreGetBigDecimal) -> Vec<BigDecimal> {
-    if let Some(pool_tvl) = pool_tvl_store.get_last(StoreKey::pool_tvl_key(&pool.address)) {
-        if pool_tvl == BigDecimal::zero() {
-            vec![BigDecimal::zero(); pool.input_tokens.len()]
-        } else {
-            pool.input_tokens
-                .iter()
-                .map(|token| {
-                    pool_tvl_store
-                        .get_last(StoreKey::pool_token_tvl_key(&pool.address, &token.address))
-                        .map_or(BigDecimal::zero(), |token_tvl| {
-                            token_tvl.div(&pool_tvl).mul(BigDecimal::from(100))
-                        })
-                })
-                .collect()
-        }
-    } else {
-        vec![BigDecimal::zero(); pool.input_tokens.len()]
-    }
 }
 
 fn update_input_token_balances(
